requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
pytest>=7.4.0
requests>=2.31.0
click>=8.1.0
tabulate>=0.9.0
